+++ tests/fhir/test_fhir_integration.py
@@ -1,16 +1,21 @@
+import pytest
 from tests.files. files import *
 from app.model.fhir.from_fhir_v1 import FromFHIRV1
 from app.model.file_handling.data_files import DataFiles
 # from uuid import UUID
 
-def _run_test(name, save=False):
+@pytest.fixture
+def anyio_backend():
+  return 'asyncio'
+
+async def _run_test(name, save=False):
   filename = f"{name}.json"
   contents = read_json(_full_path(filename))
   files = DataFiles()
   uuid = files.new()
   files.save("fhir", contents, filename)
   fhir = FromFHIRV1(uuid)
-  result = fhir.to_usdm()
+  result = await fhir.to_usdm()
   result = result.replace(uuid, 'FAKE-UUID') # UUID allocated is dynamic, make fixed
   pretty_result = json.dumps(json.loads(result), indent=2)
   result_filename = filename = f"{name}-usdm.json"
@@ -22,11 +27,14 @@ def _run_test(name, save=False):
 def _full_path(filename):
   return f"tests/test_files/fhir_v1/{filename}"
 
-def test_fhir_v1_ASP8062():
-  _run_test('ASP8062')
+@pytest.mark.anyio
+async def test_fhir_v1_ASP8062():
+  await _run_test('ASP8062', True)
 
-def test_fhir_v1_DEUCRALIP():
-  _run_test('DEUCRALIP')
+@pytest.mark.anyio
+async def test_fhir_v1_DEUCRALIP():
+  await _run_test('DEUCRALIP', True)
 
-def test_fhir_v1_IGBJ():
-  _run_test('IGBJ')
+@pytest.mark.anyio
+async def test_fhir_v1_IGBJ():
+  await _run_test('IGBJ', True)